Remove deprecated indicators (`land.freq_analysis`, `land.stats`, `land.fit`)
